modify the condition for whether wp-content folder is deleted or not #11
This ensures that we are only deleting & recreating the wp-content folder and registering the hub as a remote IF the hub is not already registered as a remote (instead of using the presence of `post_update` hook in hub).  This is because it’s possible new sites get added to the `post_update` hook that are connected to other branches for the same “hub” repository.  Keeps things indempotent.